🐛 fix: 홈화면 컨테이너 분리 로직 수정
